Dekkov-chatbot/requirements.txt
browser_use
python-dotenv
langchain_openai
accelerate
aioboto3
aiohttp

# database packages
graspologic
hnswlib
nano-vectordb
neo4j
networkx
ollama
openai
oracledb
pyvis
tenacity
# lmdeploy[all]

# LLM packages
tiktoken
torch
transformers
xxhash
lightrag-hku
